+++ src/utils/releaseNotes.js
@@ -1,6 +1,12 @@
 'use strict'
 
+const semver = require('semver')
+const _truncate = require('lodash.truncate')
+
 const md = require('markdown-it')()
+
+const MAX_PR_BODY_SIZE_LIMITATION = 65536
+const PR_BODY_TRUNCATE_SIZE = 60000
 
 function getPrNumbersFromReleaseNotes(releaseNotes) {
   const parsedReleaseNotes = md.parse(releaseNotes, {})
@@ -28,6 +34,46 @@ function getPrNumbersFromReleaseNotes(releaseNotes) {
     .filter(prNumber => Boolean(prNumber))
 
   return [...new Set(allPrNumbers)]
+}
+
+exports.getPRBody = (
+  template,
+  { newVersion, draftRelease, inputs, author, artifact }
+) => {
+  const tagsToBeUpdated = []
+  const { major, minor } = semver.parse(newVersion)
+
+  if (major !== 0) tagsToBeUpdated.push(`v${major}`)
+  if (minor !== 0) tagsToBeUpdated.push(`v${major}.${minor}`)
+
+  // Should strictly contain only non-sensitive data
+  const releaseMeta = {
+    id: draftRelease.id,
+    version: newVersion,
+    npmTag: inputs['npm-tag'],
+    opticUrl: inputs['optic-url'],
+  }
+
+  const prBody = template({
+    releaseMeta,
+    draftRelease,
+    tagsToUpdate: tagsToBeUpdated.join(', '),
+    npmPublish: !!inputs['npm-token'],
+    artifact,
+    syncTags: /true/i.test(inputs['sync-semver-tags']),
+    author,
+  })
+
+  if (prBody.length > MAX_PR_BODY_SIZE_LIMITATION) {
+    const omissionText =
+      '. *Note: Part of the release notes have been omitted from this message, as the content exceeds the size limit*'
+    return _truncate(prBody, {
+      length: PR_BODY_TRUNCATE_SIZE,
+      omission: omissionText,
+    })
+  }
+
+  return prBody
 }
 
 exports.getPrNumbersFromReleaseNotes = getPrNumbersFromReleaseNotes
